refactor: remove old auth store
