+++ apps/core/src/pages/landing.tsx
@@ -11,7 +11,7 @@ import {
   ModalTitle,
   ModalFooter
 } from '@RFWebApp/ui';
-import { useAuthStore } from '../../../../src/store/auth';
+import { useAuthStore } from '@lib/store';
 
 interface Funcionario {
   id: string;
