+++ lib/useRequireAuth.ts
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 import { useIsAuthenticated } from '@azure/msal-react';
-import { useAuthStore } from '../src/store/auth';
+import { useAuthStore } from './store';
 export function useRequireAuth() {
   const isAuthenticated = useIsAuthenticated();
   const { employeeNumber } = useAuthStore();
